+++ config/dravion.php
@@ -1,7 +1,7 @@
-﻿<?php
+<?php
 
 return [
-    'version' => '1.13.4',
+    'version' => '1.14.0',
 
     'license_server' => env('DRAVION_LICENSE_SERVER', 'https://apsbg.com/dravion-server'),
     'updates_server' => env('DRAVION_UPDATES_SERVER', 'https://apsbg.com/dravion-server'),